// types/inventoryVoucher/OrderBookingTypes.d.ts
export interface OrderBookingTableProps {
  loading: boolean;
  handleShowDeleteDialog: (id: number) => void;
  showDeleteDialog: boolean;
  setShowDeleteDialog: Dispatch<SetStateAction<boolean>>;
  setTempDeleteId: Dispatch<SetStateAction<number | null>>;
  handleDeleteOrder: () => void;
  deleteOrderBookingLoading: boolean;
  currentPage: number;
  setCurrentPage: Dispatch<SetStateAction<number>>;
  lastPage: number;
}

export interface DesignModalProps {
  form: UseFormReturn<OrderBookingFormData>;
  showDesignDialog: boolean;
  setShowDesignDialog: Dispatch<SetStateAction<boolean>>;
  handleAddDesign: () => void;
  setDesignInput: Dispatch<SetStateAction<string>>;
}

export interface OrderBookingTableData {
  Id: number;
  Order_Date: string;
  Order_No: string;
  Party_Name: string;
  Party_Id: number;
  Total_Order: string;
  Order_Status: string;
  DesignRow: {
    Design_Id: number;
    Design_Name: string;
    Design_No: string;
    Order_Qnty: string;
    Design_Rate: string;
    Wt: string;
    Wt_Rate: string;
    Tot_Wt: string;
    Polish: string;
    Tot_Polish: string;
    Image: string;
    ItemRow: {
      Item_Id: number;
      Item_Name: string;
      Item_Qnty: string;
      Item_Rate: string;
      Making_Rate: string;
      Item_Tot: string;
    }[];
  }[];
}